Move TestLINQ code to TestAddress
